% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_deviation_plot.R
\name{create_deviation_plot}
\alias{create_deviation_plot}
\title{create_deviation_plot}
\usage{
create_deviation_plot(data_ptds, coded_descriptions)
}
\value{

}
\description{
create_deviation_plot
}
\author{
Aaron Conway
}
