% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fasting_duration_ptds_plot.R
\name{create_fasting_duration_ptds_plot}
\alias{create_fasting_duration_ptds_plot}
\title{}
\usage{
create_fasting_duration_ptds_plot(data = data_ptds, fluids_corr, food_corr)
}
\value{

}
\description{

}
\author{
Aaron Conway
}
