% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fluids_corr.r
\name{create_fluids_corr}
\alias{create_fluids_corr}
\title{Correlation ptds and fluids fasting}
\usage{
create_fluids_corr(data_ptds)
}
\value{

}
\description{
Correlation ptds and fluids fasting
}
\author{
Aaron Conway
}
