% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_food_corr.r
\name{create_food_corr}
\alias{create_food_corr}
\title{Correlation ptds and food fasting}
\usage{
create_food_corr(data_ptds)
}
\value{

}
\description{
Correlation ptds and food fasting
}
\author{
Aaron Conway
}
