% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_global_thirst_corr.R
\name{create_global_thirst_corr}
\alias{create_global_thirst_corr}
\title{Correlation ptds and global thirst}
\usage{
create_global_thirst_corr(data_ptds)
}
\arguments{
\item{data_ptds}{dataframe}
}
\value{

}
\description{
Correlation ptds and global thirst
}
\author{
Aaron Conway
}
