% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_monotonicity_df_ptds6_table.R
\name{create_monotonicity_df_ptds6_table}
\alias{create_monotonicity_df_ptds6_table}
\title{}
\usage{
create_monotonicity_df_ptds6_table(monotonicity_df_ptds6)
}
\value{

}
\description{

}
\author{
Aaron Conway
}
