% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_data.R
\name{label_data}
\alias{label_data}
\title{Label data}
\usage{
label_data(data)
}
\arguments{
\item{data}{dataframe from csv downloaded from redcap}
}
\value{

}
\description{
Label data
}
\author{
Aaron Conway
}
