% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_corr_plot.R
\name{make_corr_plot}
\alias{make_corr_plot}
\title{Make a correlation plot}
\usage{
make_corr_plot(ptds7)
}
\arguments{
\item{ptds7}{data frame}
}
\value{

}
\description{
Make a correlation plot
}
\author{
Aaron Conway
}
