% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moscales_lowerbounds.R
\name{moscales_lowerbounds}
\alias{moscales_lowerbounds}
\title{Bounds}
\usage{
moscales_lowerbounds(x, lowerbounds = seq(from = 0.05, to = 0.8, by = 0.05))
}
\arguments{
\item{x}{data frame}
}
\value{

}
\description{
Bounds
}
