% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_df.R
\name{preprocess_df}
\alias{preprocess_df}
\title{Preprocessing}
\usage{
preprocess_df(data_labelled)
}
\arguments{
\item{data_labelled}{labelled data from redcap}
}
\value{

}
\description{
Preprocessing
}
\author{
Aaron Conway
}
